Clean up the DB queries a bit
