+++ apiV2/app/db/impl/query/apiv2/AuthQueries.scala
@@ -1,0 +1,66 @@
+package db.impl.query.apiv2
+
+import controllers.sugar.Requests.ApiAuthInfo
+import ore.db.DbRef
+import ore.models.api.ApiKey
+import ore.models.user.User
+import ore.permission.Permission
+
+import doobie._
+import doobie.implicits._
+import doobie.postgres.implicits._
+import doobie.postgres.circe.jsonb.implicits._
+import doobie.implicits.javasql._
+import doobie.implicits.javatime.JavaTimeLocalDateMeta
+
+object AuthQueries extends APIV2Queries {
+
+  def getApiAuthInfo(token: String): Query0[ApiAuthInfo] = {
+    sql"""|SELECT u.id,
+          |       u.created_at,
+          |       u.full_name,
+          |       u.name,
+          |       u.email,
+          |       u.tagline,
+          |       u.join_date,
+          |       u.read_prompts,
+          |       u.language,
+          |       ak.name,
+          |       ak.owner_id,
+          |       ak.token,
+          |       ak.raw_key_permissions,
+          |       aks.token,
+          |       aks.expires,
+          |       CASE
+          |           WHEN u.id IS NULL THEN 1::BIT(64)
+          |           ELSE (coalesce(gt.permission, B'0'::BIT(64)) | (
+          |               -- Default permissions if no global trust is found
+          |                       1::BIT(64) | -- ViewPublicInfo
+          |                       (1::BIT(64) << 1) | -- Permission.EditOwnUserSettings
+          |                       (1::BIT(64) << 2) -- Permission.EditApiKeys
+          |               )) & coalesce(ak.raw_key_permissions, (-1)::BIT(64))
+          |           END
+          |    FROM api_sessions aks
+          |             LEFT JOIN api_keys ak ON aks.key_id = ak.id
+          |             LEFT JOIN users u ON aks.user_id = u.id
+          |             LEFT JOIN global_trust gt ON gt.user_id = u.id
+          |    WHERE aks.token = $token""".stripMargin.query[ApiAuthInfo]
+  }
+
+  def findApiKey(identifier: String, token: String): Query0[(DbRef[ApiKey], DbRef[User])] =
+    sql"""SELECT k.id, k.owner_id FROM api_keys k WHERE k.token_identifier = $identifier AND k.token = crypt($token, k.token)"""
+      .query[(DbRef[ApiKey], DbRef[User])]
+
+  def createApiKey(
+      name: String,
+      ownerId: DbRef[User],
+      tokenIdentifier: String,
+      token: String,
+      perms: Permission
+  ): doobie.Update0 =
+    sql"""|INSERT INTO api_keys (created_at, name, owner_id, token_identifier, token, raw_key_permissions)
+          |VALUES (now(), $name, $ownerId, $tokenIdentifier, crypt($token, gen_salt('bf')), $perms)""".stripMargin.update
+
+  def deleteApiKey(name: String, ownerId: DbRef[User]): doobie.Update0 =
+    sql"""DELETE FROM api_keys k WHERE k.name = $name AND k.owner_id = $ownerId""".update
+}
