+++ ore/test/db/apiv2/PageQueriesSpec.scala
@@ -1,0 +1,31 @@
+package db.apiv2
+
+import controllers.apiv2.Pages
+import db.DbSpec
+import db.impl.query.apiv2.PageQueries
+import ore.OreConfig
+
+import org.junit.runner.RunWith
+import org.scalatestplus.junit.JUnitRunner
+import pureconfig.ConfigSource
+
+@RunWith(classOf[JUnitRunner])
+class PageQueriesSpec extends DbSpec {
+
+  implicit val config: OreConfig = ConfigSource.default.loadOrThrow[OreConfig]
+
+  test("getPage") {
+    check(PageQueries.getPage("foo", "bar", "baz"))
+  }
+
+  test("pageList") {
+    check(PageQueries.pageList("foo", "bar"))
+  }
+
+  test("patchPage") {
+    check(
+      PageQueries
+        .patchPage(Pages.PatchPageF[Option](Some("foo"), Some(Some("bar")), None), Some("baz"), 5L, Some(Some(5L)))
+    )
+  }
+}
